Chains strong hashes on weak hash collisions.
Earlier, whenever the weak hash collides we overwrite the value entry of in the HashMap of hashes. I fixed this by storing the HashObjects as an ArrayList in the HashMap and simply appending to it whenever we have a collision. A function to search for a strongHash inside this ArrayList was also written.