Fixes the rolling and running Adler32 computations
There were a number of problems related to computing both the aggregate
hash and rolling version of the checksum. This resulted in the rolling
checksum eventually breaking our checksum and effectively ruining our
ability to match anymore blocks which then results in sending over the
remaining literal data. Most of the issues dealt with bitwise arithmetic
as well as flaws in our interpretation of the equation.